Extract out pagination from helpers
