+++ controllers/apply-next/form-router.js
@@ -1,22 +1,32 @@
 'use strict';
-const { concat, head, isEmpty, omit, get, set, unset } = require('lodash');
+const { get, set, unset, concat, flatMap, head, isEmpty, omit } = require('lodash');
 const express = require('express');
 const path = require('path');
 const Raven = require('raven');
 
+const applicationsService = require('../../services/applications');
 const cached = require('../../middleware/cached');
 const { requireUserAuth } = require('../../middleware/authed');
-const applicationsService = require('../../services/applications');
-
 const { normaliseErrors } = require('../../modules/errors');
-const {
-    FORM_STATES,
-    calculateFormProgress,
-    enhanceForm,
-    fieldsForStep,
-    mapFields,
-    nextAndPrevious
-} = require('./helpers');
+
+const { FORM_STATES, calculateFormProgress, enhanceForm } = require('./helpers');
+const { nextAndPrevious } = require('./lib/pagination');
+
+function mapFields(fields) {
+    return {
+        fields: fields,
+        names: fields.map(field => field.name),
+        messages: fields.reduce((obj, field) => {
+            obj[field.name] = field.messages;
+            return obj;
+        }, {})
+    };
+}
+
+function fieldsForStep(step) {
+    const fields = flatMap(step.fieldsets, 'fields');
+    return mapFields(fields);
+}
 
 function initFormRouter(formModel) {
     const router = express.Router();
